Create new EditProfile component that handles the form to edit profile
